#2 Update items for Blood DPS
The commit made updates to the specific item ranks for the Death Knight specialization in Bistooltip_wh_items. Spec_name: "Blood dps" was the main focus. Darkmoon Greatness was switched out for Death's Verdict in PR WH, to be consistent with Frost. And an Error from the WowHead site, where a Set Piece Shoulder was placed in the BiS List for chests, was switched the corresponding chest piece